pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "financial-news"
version = "1.0.0"
description = "Self-hosted financial market intelligence system"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # Database
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "alembic>=1.13",

    # Task queue
    "celery[redis]>=5.4",
    "redis>=5.0",

    # Ingestion
    "httpx>=0.27",
    "feedparser>=6.0",
    "trafilatura>=1.9",
    "python-dateutil>=2.9",
    "pyyaml>=6.0",

    # RAG / embeddings
    "sentence-transformers>=3.0",
    "numpy>=1.26",

    # Logging
    "structlog>=24.1",

    # Email
    # (imaplib + smtplib are stdlib)

    # Optional: backtesting price data
    # "yfinance>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "httpx>=0.27",
    "anyio>=4.3",
    "aiosqlite>=0.20",
]

[project.scripts]
fnews-backtest = "packages.backtest.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "packages*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
